% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{get_wildcards_tbl}
\alias{get_wildcards_tbl}
\title{Get the Snakemake wildcards as a tibble}
\usage{
get_wildcards_tbl()
}
\value{
a tibble of wildcards, with columns as names and rows as values
}
\description{
Get the Snakemake wildcards as a tibble
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
