% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_helpers.R
\name{is_nearly_whole}
\alias{is_nearly_whole}
\title{Checks whether a number is near to a whole number}
\usage{
is_nearly_whole(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks whether a number is near to a whole number
}
\examples{
is_nearly_whole(.Machine$double.eps^0.5)
is_nearly_whole(.Machine$double.eps^0.6)
is_nearly_whole(1)
}
