% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_deps}
\alias{load_deps}
\title{Install & load packages}
\usage{
load_deps(...)
}
\arguments{
\item{...}{package names to install & load}
}
\description{
Install & load packages
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
