% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{log_snakemake}
\alias{log_snakemake}
\title{Save output, messages, warnings, and errors to the Snakemake log file}
\usage{
log_snakemake(quiet = TRUE)
}
\arguments{
\item{quiet}{Silence messages about the status of the snakemake object and
log file (default: \code{TRUE}).}
}
\description{
This function checks whether a log file was specified in the Snakemake rule.
If so, it directs any output, messages, warnings, or errors to the
rule-specific log file.
See the Snakemake documentation on \href{https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#log-files}{log files}
and \href{https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#r-and-r-markdown}{R scripts}
for more details.
}
\examples{
# The Snakemake object doesn't exist, so nothing happens
log_snakemake(quiet = FALSE)
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
