% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tax.R
\name{parse_tax}
\alias{parse_tax}
\title{Convert taxonomy strings into dataframe of labels based on taxonomic classification}
\usage{
parse_tax(dat)
}
\arguments{
\item{dat}{dataframe from mothur taxonomy file with columns \code{OTU}, \code{Size}, and \code{Taxonomy}}
}
\value{
a wide dataframe with taxonomic labels
}
\description{
Convert taxonomy strings into dataframe of labels based on taxonomic classification
}
\examples{
taxonomy_filepath <- system.file("extdata",
  "test.taxonomy",
  package = "schtools"
)
taxonomy_tbl <- read_tax(taxonomy_filepath)
head(taxonomy_tbl)
}
\author{
Nick Lesniak, \email{nlesniak@umich.edu}
}
