% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dist.R
\name{read_dist}
\alias{read_dist}
\title{Read in lower left triangular matrix from file}
\usage{
read_dist(dist_filename)
}
\arguments{
\item{dist_filename}{filename of lower left triangular matrix (.dist)}
}
\value{
distance matrix as a tibble
}
\description{
Read in lower left triangular matrix from file
}
\examples{
# TODO: add dist file to inst/ext-data and create a real example of using this
}
\author{
Nick Lesniak, \email{nlesniak@umich.edu}
}
