% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schtools.R
\docType{package}
\name{schtools-package}
\alias{schtools}
\alias{schtools-package}
\title{schtools: Schloss Lab Tools for Reproducible Microbiome Research}
\description{
A collection of useful functions and example code created and used by the
Schloss Lab for reproducible microbiome research. Perform common tasks like
read files created by \href{https://mothur.org/}{mothur}, tidy up your microbiome
data, and format rmarkdown documents for publication. See the
\href{http://www.schlosslab.org/schtools/}{website} for more information,
documentation, and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.schlosslab.org/schtools/}
  \item \url{https://github.com/SchlossLab/schtools}
  \item Report bugs at \url{https://github.com/SchlossLab/schtools/issues}
}

}
\author{
\strong{Maintainer}: Kelly Sovacool \email{sovacool@umich.edu} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})

Authors:
\itemize{
  \item Nicholas Lesniak \email{nlesniak@umich.edu} (\href{https://orcid.org/0000-0001-9359-5194}{ORCID})
  \item Sarah Lucas (\href{https://orcid.org/0000-0003-1676-5801}{ORCID})
  \item Courtney Armour (\href{https://orcid.org/0000-0002-5250-1224}{ORCID})
  \item Patrick Schloss \email{pschloss@umich.edu}
}

Other contributors:
\itemize{
  \item Jacqueline Moltzau (\href{https://orcid.org/0000-0003-1398-5980}{ORCID}) [contributor]
  \item Andrew Hannah [contributor]
  \item Nielson Baxter [contributor]
  \item Alyxandria Schubert [contributor]
  \item Kathryn Iverson [contributor]
}

}
\keyword{internal}
