% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_helpers.R
\name{set_knitr_opts}
\alias{set_knitr_opts}
\title{Set knitr chunk options & inline hook}
\usage{
set_knitr_opts()
}
\description{
Call this function in the setup chunk of your R Markdown files.
}
\author{
Pat Schloss \email{pschloss@umich.edu}

Kelly Sovacool \email{sovacool@umich.edu}
}
