% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_sovacool}
\alias{theme_sovacool}
\title{Kelly's go-to theme for ggplot2}
\usage{
theme_sovacool()
}
\value{
list of ggproto objects
}
\description{
Uses \code{ggplot2::theme_bw()} and removes margins.
}
\examples{
library(ggplot2)
ggplot(mtcars) +
  aes(x = mpg, y = wt, color = cyl) +
  geom_point() +
  theme_sovacool()
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
