% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_helpers.R
\name{inline_hook}
\alias{inline_hook}
\title{Inline hook for knitr to paste human-readable numbers.}
\usage{
inline_hook(x)
}
\arguments{
\item{x}{inline code}
}
\value{
formatted \code{x} if numeric, otherwise \code{x} unmodified.
}
\description{
Pastes formatted \code{x} if numeric, otherwise \code{x} unmodified.
Circumvents R's automatic scientific notation.
}
\examples{
inline_hook(0.02)
inline_hook(.Machine$double.eps^0.5)
inline_hook("this is a string")
}
\author{
Pat Schloss \email{pschloss@umich.edu}

Kelly Sovacool \email{sovacool@umich.edu}
}
