% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dist.R
\name{read_dist}
\alias{read_dist}
\title{Read in lower left triangular matrix from file}
\usage{
read_dist(dist_filename)
}
\arguments{
\item{dist_filename}{filename of lower left triangular matrix (.dist)}
}
\value{
distance matrix as a tibble
}
\description{
Read in lower left triangular matrix from file
}
\examples{
dist_filepath <- system.file("extdata",
                             "sample.final.thetayc.0.03.lt.ave.dist",
                             package = "mothuR")
dist_tbl <- read_dist(dist_filepath)
head(dist_tbl)
}
\author{
Nick Lesniak, \email{nlesniak@umich.edu}
}
