% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tax.R
\name{read_tax}
\alias{read_tax}
\title{Convert taxonomy strings into dataframe of labels based on taxnomic classification}
\usage{
read_tax(taxonomy_filename)
}
\arguments{
\item{taxonomy_filename}{filename of taxonomy file}
}
\value{
dataframe of taxonomic labels
}
\description{
Convert taxonomy strings into dataframe of labels based on taxnomic classification
}
\examples{
taxonomy_filepath <- system.file("extdata",
  "test.taxonomy",
  package = "schtools"
)
taxonomy_tbl <- read_tax(taxonomy_filepath)
head(taxonomy_tbl)
}
\author{
Nick Lesniak, \email{nlesniak@umich.edu}
}
