% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schtools.R
\docType{package}
\name{schtools}
\alias{schtools}
\title{schtools: Schloss Lab Tools for Reproducible Microbiome Research}
\description{
A collection of useful functions and example code created and used by the
Schloss Lab for reproducible microbiome research. Perform common tasks like
read mothur output files, tidy up your microbiome data, and format rmarkdown
documents for publication. See the website
\url{http://www.schlosslab.org/schtools/} for more information,
documentation, and examples.
}
